update benchmark and localisation
I have connected localization.py. But because you output 3 images in image.py, which is a bit small for the cluster, I changed the output to 6 images, and then output each image once, so one image will have 6 points.

Then I rewrote benchmark.py, because its original logic was not what we wanted at all, and it seemed that match was re-run (a duplicate of image.py), and it took a very long time to run.

My design logic is as follows: among the n results output by imgae.py (my setting: 6, because of the previous localization.py), any picture that matches our validation will count as one point. If you match two, it counts as two points, and so on.